+++ CHANGELOG.md
@@ -7,3 +7,9 @@
     - showKDialogContent
     - showKDialogConfirm
     - showKDialogProcessing
+
+## 0.0.1
+
+- Fix showKDialogConfirm and showKBottomErrorMessage response
+    - response false by default
+- Fix showKDialogConfirm width when message is short